A lot of effort, not a lot of productivity.
Mostly because I'm kind of tired, and the expected amount of work is
making me quail; I need to rest.

Biggest change has happened in the background, in game.py; we've created
an Actor class and move the 2-d matrix that's supposed to be their
memory in there. Nobody references it yet, it still compiles, but
yeah. Um.

Next change is that in stead of loading an empty grid, we have a new map
file (in txt/empty-chamber.map) that gets loaded into the Level
class. Not getting referenced yet.

Other than that it's mostly cleanup. Next round: dig into viewport.py,
get it to reference the level without breakage, which should really
involve *only* figuring out how to cut a rectangular-sized chunk of the
level map, rendering that, and moving it. Not to hard. Ha, ha, ha.

No, seriously, it sounds like fun but I'm too tired.