Success! Finally!
After... uh... literally several days of banging my head on the code,
over 100 lines of shoddy crap, and FINALLY, we get a little sprite walking
about on an empty chamber.

Granted, the sprite doesn't know what walls are, and the code could
REALLY use a MAJOR cleanup and I don't even know what else I've done to
get here, but THERE YOU HAVE IT. It's rendering. Oh god.

Let me see...

There's a lot of work to jam through an Actor() class that has its own
memory, to also make sure that all the Level() does to pass to Game() is
JUST what it needs to render, and NOTHING else. There's BASIC bounds
checking, if only so that the player doesn't walk off the level, but...

...

That's really it.

...

T__T